requirements.txt
streamlit==1.2.0
google-cloud-vision==2.4.2